complete refactoring of kernels
