+++ diffhtwo/experimental/utils.py
@@ -8,8 +8,8 @@ from dsps.cosmology.defaults import DEFAULT_COSMOLOGY
 from jax import jit as jjit
 from jax.tree_util import tree_flatten_with_path
 
+from .kernels.spec_kern import n_spec_q_ms_burst
 from .lightcone_generators import generate_lc_data
-from .n_specphot import n_spec_q_ms_burst
 
 
 @jjit
