+++ scripts/config.yaml
@@ -4,7 +4,7 @@ start_runid: "run50"
 start_fit_type: "spspop+diffstarpop"
 
 fit_runid: "run51"
-fit_type: "diffstarpop"
+fit_type: "diffstarpop+spspop+ssperrpop"
 
 epoch:
   n_it: 1
